feat: split the library into modules and uses pub use to expose a small public API
