+++ Cargo.toml
@@ -4,4 +4,3 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-coerce_pattern = "0.1.0"
